Socket auxiliar functions moved shared file
